[#161] Restructuring Template
