+++ Server/src/controllers/dataUploadController.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { DataProcessService } from '../services/dataProcess/dataProcessService'
+import { DataProcessService } from '../services/dataProcess/DataProcessService'
 
 /**
  * The dataUploadController is responsible for processing providing instructions to the application if a request comes in
